feat(base-types): reorganize types for runtime logging
