+++ packages/base-types/src/runtimeLogs/index.ts
@@ -1,2 +1,3 @@
 export * from './logs';
 export * from './runtime';
+export { Iso8601Timestamp } from './utils';
