+++ packages/base-types/src/runtimeLogs/logs/levels.ts
@@ -1,0 +1,6 @@
+export enum LogLevel {
+  ERROR = 'error',
+  WARN = 'warn',
+  INFO = 'info',
+  VERBOSE = 'verbose',
+}
