+++ packages/base-types/src/runtimeLogs/logs/steps/capture.ts
@@ -1,9 +1,9 @@
-import { LoggingNodeType } from '../../utils';
 import { BaseStepLog } from '../base';
+import { StepLogKind } from '../kinds';
 
 // TODO: Make sure there isn't other information that should be included here
 export type CaptureStepLog = BaseStepLog<
-  LoggingNodeType.CAPTURE,
+  StepLogKind.CAPTURE,
   {
     slots: Array<{
       utteredValue: string;
