+++ packages/base-types/src/runtimeLogs/logs/steps/condition.ts
@@ -1,8 +1,9 @@
-import { LoggingNodeType, PathReference } from '../../utils';
+import { PathReference } from '../../utils';
 import { BaseStepLog } from '../base';
+import { StepLogKind } from '../kinds';
 
 export type ConditionStepLog = BaseStepLog<
-  LoggingNodeType.CONDITION,
+  StepLogKind.CONDITION,
   {
     path: PathReference;
   }
