+++ packages/base-types/src/runtimeLogs/logs/steps/customAction.ts
@@ -1,10 +1,11 @@
 import { Nullable } from '@voiceflow/common';
 
-import { LoggingNodeType, PathReference } from '../../utils';
+import { PathReference } from '../../utils';
 import { BaseStepLog } from '../base';
+import { StepLogKind } from '../kinds';
 
 export type CustomActionStepLog = BaseStepLog<
-  LoggingNodeType.CUSTOM_ACTION,
+  StepLogKind.CUSTOM_ACTION,
   {
     actionBody: Nullable<string>;
     path: PathReference;
