+++ packages/base-types/src/runtimeLogs/logs/steps/exit.ts
@@ -1,12 +1,14 @@
 import { EmptyObject } from '@voiceflow/common';
 
-import { LoggingNodeType } from '../../utils';
 import { BaseStepLog } from '../base';
+import { StepLogKind } from '../kinds';
+import { LogLevel } from '../levels';
 
 export type ExitStepLog = BaseStepLog<
-  LoggingNodeType.EXIT,
+  StepLogKind.EXIT,
   {
     /** The state of the program on exit. */
     state: EmptyObject;
-  }
+  },
+  LogLevel.VERBOSE
 >;
