+++ packages/base-types/src/runtimeLogs/logs/steps/speak.ts
@@ -1,8 +1,8 @@
-import { LoggingNodeType } from '../../utils';
 import { BaseStepLog } from '../base';
+import { StepLogKind } from '../kinds';
 
 export type SpeakStepLog = BaseStepLog<
-  LoggingNodeType.SPEAK,
+  StepLogKind.SPEAK,
   {
     text: string;
   }
