+++ packages/base-types/src/runtimeLogs/logs/steps/start.ts
@@ -1,7 +1,7 @@
 import { EmptyObject } from '@voiceflow/common';
 
-import { LoggingNodeType } from '../../utils';
 import { BaseStepLog } from '../base';
+import { StepLogKind } from '../kinds';
 
 // TODO: Consider including information from the global.conversation_start event here
-export type StartStepLog = BaseStepLog<LoggingNodeType.START, EmptyObject>;
+export type StartStepLog = BaseStepLog<StepLogKind.START, EmptyObject>;
